requirements.txt
boto3==1.42.60
botocore==1.42.60
jmespath==1.0.1
python-dotenv==1.2.2
s3transfer==0.16.0
urllib3==2.6.3
six==1.17.0
